% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_q.R
\name{check_q}
\alias{check_q}
\title{Is the ACN tumor profile q, corresponding to RCN sample profile r, tumor purity, tumor ploidy non-negative in every segment?}
\usage{
check_q(pu_plo, r)
}
\arguments{
\item{pu_plo}{vector with two components; tumor purity, tumor ploidy}

\item{r}{sample RCN profile (vector)}
}
\value{
logical; True, if tumor ACN profile q is non-negative in every segment; False otherwise
}
\description{
\code{check_q} checks whether tumor ACN profile q is non-negative.
}
