% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_cnhplus.R
\name{find_cnhplus}
\alias{find_cnhplus}
\title{For sample RCN profile r finds CNH+ and the associated tumor purity, tumor ploidy}
\usage{
find_cnhplus(grid, r, w, k, plus = TRUE)
}
\arguments{
\item{grid, }{grid of purities, ploidies (matrix)}

\item{r}{sample RCN profile (vector)}

\item{w}{vector of segment widths}

\item{k}{how many candidate pairs of purity, ploidy for the solution to return}

\item{plus}{if True (default) CNH+ is solved for; if False, CNH}
}
\value{
top k elements of the grid together with the associated values of the objective function kappa
}
\description{
\code{find_cnhplus} by searching over grid, finds value of CNH+ and associated tumor purity, tumor ploidy for a sample RCN.
It also makes possible to find CNH.
}
\seealso{
\code{plot_profile}
}
