% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappa.R
\name{kappa}
\alias{kappa}
\title{Computes value of the objective function kappa for purity, ploidy pair, RCN sample profile r and associated segments width vector w}
\usage{
kappa(pu_plo, r, w)
}
\arguments{
\item{pu_plo}{vector with two components: tumor purity, tumor ploidy}

\item{r}{sample RCN profile}

\item{w}{vector of segments width}
}
\value{
value of kappa
}
\description{
\code{kappa} evaluates the objective function kappa.
}
