% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappa_q.R
\name{kappa_q}
\alias{kappa_q}
\title{Computes value of the objective function kappa for ACN profile q and associated segments width vector w}
\usage{
kappa_q(q, w)
}
\arguments{
\item{q}{ACN profile of tumor (vector)}

\item{w}{vector of segments width}
}
\value{
value of kappa
}
\description{
\code{kappa_q} evaluates the objective function kappa.
}
