% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_grid_plus.R
\name{make_grid_plus}
\alias{make_grid_plus}
\title{For given RCN profile r of sample and a grid of purities and ploidies, those pairs for which the tumor ACN profile q fails the non-negativity requirement are removed from the grid}
\usage{
make_grid_plus(grid, r)
}
\arguments{
\item{grid}{grid of purities, ploidies (matrix)}

\item{r}{sample RCN  profile (vector)}
}
\value{
grid (matrix) of purities, ploidies
}
\description{
\code{make_grid_plus} removes from the grid all pairs of (purity, ploidy) for which the tumor ACN profile q has at least one segment's copy negative; tumor ACN profile q is obtained from the sample RCN profile q and the tumor purity, the tumor ploidy.
}
