% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_survival.R
\name{plot_survival}
\alias{plot_survival}
\title{Plot of survival functions of samples from TCGA study with CNH below or equal to median vs samples with CNH above median}
\usage{
plot_survival(
  study_name,
  survival_time,
  survival_event,
  cnh,
  type,
  ylim = c(0, 1)
)
}
\arguments{
\item{study_name}{name of TCGA study}

\item{survival_time}{vector of survival times of subjects in the study}

\item{survival_event}{vector of survival events for subjects in the study}

\item{cnh}{vector of CNH values}

\item{type}{string to be placed in legend (e.g., 'unfiltered CNH+')}

\item{ylim}{permits to limit the y axis in survival plot}
}
\value{
ggsurvplot
}
\description{
\code{plot_survival} uses survival data for TCGA study and values of CNH to create plot
exhibiting Kaplan Meier survival curves for two groups of samples:
those with CNH below median and those with CNH above median. P-value for the log-rank test
of equality of the survival curves is returned in the plot together with the number at risk table.
}
\seealso{
\code{analyze_TCGA_study}
}
