% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2q.R
\name{r2q}
\alias{r2q}
\title{Computes Absolute Copy Number (ACN) tumor profile from Relative Copy Number (RCN) sample profile and tumor purity, tumor ploidy}
\usage{
r2q(r, purity, ploidy)
}
\arguments{
\item{r}{sample RCN}

\item{purity}{tumor purity (scalar)}

\item{ploidy}{tumor ploidy (scalar)}
}
\value{
ACN tumor profile (vector)
}
\description{
\code{r2q} computes ACN of tumor for tumor RCN, tumor purity, tumor ploidy.
}
