% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topk.R
\name{topk}
\alias{topk}
\title{Grid of tumor purities, tumor ploidies is sorted by the associated values of kappas (values of the objective function)
and top k values (i.e., with the smallest kappa) are returned}
\usage{
topk(grid, kappas, k)
}
\arguments{
\item{grid}{grid of tumor purities, tumor ploidies (matrix)}

\item{kappas}{vector of associated values of the objective function kappa}

\item{k}{top k to be returned}
}
\value{
elements of the grid with the k smallest values of kappa, together with the corresponding valaues of kappa
}
\description{
Grid of tumor purities, tumor ploidies is sorted by the associated values of kappas (values of the objective function)
and top k values (i.e., with the smallest kappa) are returned
}
\seealso{
\code{find_cnhplus}
}
